Separate cip25 from cip68
